arreglo guardar_archivos con gaufrette

+++ carrusel/contenidoconfig.php
@@ -20,7 +20,7 @@ echo(estilo_bootstrap());
 <div class="container">
 		<h5>CONFIGURACI&Oacute;N DE CARRUSEL Y CONTENIDOS RELACIONADOS</h5>
 		<br/>
-		
+
 
 <?php
 $campos = array(
@@ -69,13 +69,13 @@ include_once("../footer.php");
 function pintar_formulario($accion) {
     global $ruta_db_superior, $conn;
     if (isset($_REQUEST["id"]) && $_REQUEST["id"]) {
-     $contenido=busca_filtro_tabla("contenidos_carrusel.*,".fecha_db_obtener('fecha_inicio','Y-m-d')." as fecha_inicio,".fecha_db_obtener('fecha_fin','Y-m-d')." as fecha_fin","contenidos_carrusel","idcontenidos_carrusel=".$_REQUEST["id"],"",$conn); 
+     $contenido=busca_filtro_tabla("contenidos_carrusel.*,".fecha_db_obtener('fecha_inicio','Y-m-d')." as fecha_inicio,".fecha_db_obtener('fecha_fin','Y-m-d')." as fecha_fin","contenidos_carrusel","idcontenidos_carrusel=".$_REQUEST["id"],"",$conn);
     }
    include_once("../calendario/calendario.php");
    ?>
 <script type="text/javascript" src="../js/jquery.js"></script>
 <script type="text/javascript" src="../js/jquery.validate.js"></script>
-<script type="text/javascript" src="../js/jquery.spin.js"></script>    
+<script type="text/javascript" src="../js/jquery.spin.js"></script>
 <script type="text/javascript" src="<?php echo $ruta_db_superior; ?>tinymce34/jscripts/tiny_mce/tiny_mce.js"></script>
 <script src="<?php echo $ruta_db_superior;?>dropzone/dist/dropzone.js"></script>
 
@@ -114,7 +114,7 @@ width:"350px"
     		submitHandler: function(form) {
 				<?php encriptar_sqli("form1",0,"form_info",$ruta_db_superior);?>
 			    form.submit();
-			    
+
 			  }
     	});
     	$.spin.imageBasePath = '../images/';
@@ -128,28 +128,28 @@ width:"350px"
         '/',
         ['Font','FontSize'],['Bold','Italic','Underline','Strike'],
         ['JustifyLeft','JustifyCenter','JustifyRight','JustifyBlock'],
-        ['NumberedList','BulletedList','-','Outdent','Indent','Blockquote']        
+        ['NumberedList','BulletedList','-','Outdent','Indent','Blockquote']
     ];
     </script>
-    
+
 		<ul class="nav nav-tabs">
-		
+
 			 <li ><a href='sliderconfig.php'>Inicio</a ></li>
 		<?php if($accion=='adicionar') { ?>
-					
+
 				 <li><a href='sliderconfig.php?accion=adicionar'>Adicionar Carrusel</a ></li>
 				 <li  class="active"><a href='contenidoconfig.php?accion=adicionar'>Adicionar Contenido</a ></li>
-		<?php }else{ ?>	
-				
+		<?php }else{ ?>
+
 				  <li ><a href='sliderconfig.php?accion=adicionar'>Adicionar Carrusel</a ></li>
 			      <li><a href='contenidoconfig.php?accion=adicionar'>Adicionar Contenido</a ></li>
 			      <li class="active"><a href='#'>Editar Contenido</a ></li>
-		<?php } ?>			
-			
-		</ul>		
-		<br/>   
-    
-   <?php    
+		<?php } ?>
+
+		</ul>
+		<br/>
+
+   <?php
    echo "<br /><fieldset><legend>" . ucwords($accion . " contenido")."</legend></fieldset><br /><br /><form action='contenidoconfig.php' name='form1' method='post' id='form1' enctype='multipart/form-data'><table class='table table-bordered table-striped'>";
    echo "<tr><td  style='text-align: center; background-color:#57B0DE; color: #ffffff;'>NOMBRE*</td><td><input class='required'  type='text' name='nombre' value='".@$contenido[0]["nombre"]."'></td></tr>";
    echo "<tr><td  style='text-align: center; background-color:#57B0DE; color: #ffffff;'>CARRUSEL*</td><td><select class='required'  type='text' name='carrusel_idcarrusel'>";
@@ -157,7 +157,7 @@ width:"350px"
    for($i=0;$i<$carrusel["numcampos"];$i++) {
        echo "<option value='".$carrusel[$i]["idcarrusel"]."' ";
        if($carrusel[$i]["idcarrusel"]==@$contenido[0]["carrusel_idcarrusel"]) {
-         echo " selected "; 
+         echo " selected ";
         }
        echo ">".$carrusel[$i]["nombre"]."</option>";
       }
@@ -187,7 +187,7 @@ width:"350px"
       if($valor==@$contenido[0]["align"]) {
         echo " checked ";
         }
-      echo ">$nombre&nbsp;&nbsp;"; 
+      echo ">$nombre&nbsp;&nbsp;";
      }
    echo "</td></tr>";
    echo "<tr><td  style='text-align: center; background-color:#57B0DE; color: #ffffff;' width=20%>Orden*</td><td><input class='required'  type='input' name='orden' id='orden' value='".@$contenido[0]["orden"]."'></td></tr>";
@@ -257,7 +257,7 @@ width:"350px"
 function guardar_adicionar($accion, $datos, $carrusel) {
     global $ruta_db_superior, $conn;
  $nwidth=$carrusel[0]["alto"];
- $nheight=$carrusel[0]["alto"]; 
+ $nheight=$carrusel[0]["alto"];
 
     $sql1 = "insert into contenidos_carrusel(" . implode(",", array_keys($datos)) . ") values(" . implode(",", array_values($datos)) . ")";
  phpmkr_query($sql1,$conn);
@@ -293,46 +293,53 @@ function guardar_editar($accion, $datos, $carrusel) {
       }
         @unlink($ruta_db_superior . $contenido[0]["imagen"]);
     }
-  
+
     guardar_archivos($id, $_REQUEST["form_uuid"]);
  redirecciona($ruta_db_superior."carrusel/sliderconfig.php");
 }
- 
- 
+
+
 /**
  * @param id
  */
- 
+
 function guardar_archivos($id, $form_uuid) {
     global $ruta_db_superior, $conn;
-		 
-    //$archivos = busca_filtro_tabla("", "anexos_tmp", "uuid = '$form_uuid' AND idformato=$idformato", "", $conn);
-    //$archivos = busca_filtro_tabla("", "anexos_tmp", "uuid = '$form_uuid' AND idformato=$idformato", "", $conn);
+
     $archivos = busca_filtro_tabla("", "anexos_tmp", "uuid = '$form_uuid'", "", $conn);
     for ($j = 0; $j < $archivos["numcampos"]; $j++) {
         $ruta_temporal = $ruta_db_superior . $archivos[$j]["ruta"];
- 
-        if (file_exists($ruta_temporal)) {
-            $datos_anexo = pathinfo($ruta_temporal);
 
+        if (file_exists($ruta_temporal)) {
+
+            require_once $ruta_db_superior . 'StorageUtils.php';
+            require_once $ruta_db_superior . 'filesystem/SaiaStorage.php';
+
+            $datos_anexo = pathinfo($ruta_temporal);
             $extension = $datos_anexo["extension"];
-            $temp_filename = uniqid() . "." . $extension;
-	  $aux=RUTA_CARRUSEL_IMAGENES;
-            $dir_anexos = $ruta_db_superior . $aux;
-            crear_destino($dir_anexos);
-            $imagen_reducida = $dir_anexos . $temp_filename;
-            if (copy($ruta_temporal, $imagen_reducida)) {
-                $sql1 = "update contenidos_carrusel set imagen='" . $aux . $temp_filename . "' where idcontenidos_carrusel=" . $id;
- 			phpmkr_query($sql1,$conn);
+
+            $aleatorio = uniqid();
+            $tipo_almacenamiento = new SaiaStorage("imagenes");
+            $imagen_reducida = RUTA_CARRUSEL_IMAGENES . $aleatorio . "." . $extension;
+            $resultado = $tipo_almacenamiento->almacenar_recurso($imagen_reducida, $ruta_temporal, false);
+
+            $ruta_anexos = array(
+                "servidor" => $tipo_almacenamiento->get_ruta_servidor(),
+                "ruta" => $imagen_reducida
+            );
+            if ($tipo_almacenamiento->get_filesystem()->has($imagen_reducida)) {
+                $ruta_anexos = json_encode($ruta_anexos);
+                $sql1 = "update contenidos_carrusel set imagen='" . $ruta_anexos . "' where idcontenidos_carrusel=" . $id;
+                phpmkr_query($sql1, $conn);
+
                 @unlink($ruta_temporal);
                 unlink("$ruta_temporal.lock");
                 //Eliminar los pendientes de la tabla temporal
                 $sql2 = "DELETE FROM anexos_tmp WHERE idanexos_tmp = " . $archivos[$j]["idanexos_tmp"];
                 phpmkr_query($sql2) or die($sql2);
-      }
+            }
+        }
+    }
 }
-}
-}
-
 ?>
 </div>
